feat(database): Initial database implementation and CLI.
Adds an `internal/config` package for determining the default path to
the database.

Adds an `internal/database` package for interacting with the database.
Presently the database is hardcoded to use the sqlite3 driver. This
package is split into `database.go` which is used to interact with the
database file itself, and `operations.go` which contains all of the
functions that manipulate data within the database.

Database migrations are located at `internal/database/migrations` and
are embedded within the built binary. This allows the binary to run
database migrations as needed. A single initial migration is
implemented.

A command-line application is implemented in `cmd/rxm` in the file
`cmd/rxm/rxm.go`. This application is meant to mimic the desired final
form of the command-line interface, but currently the CLI does pure CRUD
operations with the database. The desired implementation performs more
calculations with the data, and will be implemented in a future
changeset.